Fix token auth
It seems that audience and scope of the authorizationParams of
the getAccessTokenSilently call need to be the same (or subset?)
of the Auth0Provider.

If not, this manifests as an Auth failure on browsers that prevent
tracking of third-party cookiee. This indicates that under the hood,
Auth0 SDK uses cookie instead of refresh token and local storage
to request access token, which causes it to fail for browsers with
Intelligent Tracking Prevention (ITP) turned on